Vuetify

+++ backend/controllers/userController.js
@@ -4,40 +4,34 @@ const fs = require('fs');
 const bcrypt = require("bcrypt");
 
 //Afficher un utilisateur
-exports.showUser = async (req, res) =>{
+exports.showUser = async (req, res) => {
     try {
         const user = await User.findOne({
             where: {
                 id: req.params.id
             },
-            attributes: ["id", "username", "profilepic", "email"]
+            attributes: ["id", "username", "profilepic", "email", "isMod"]
         });
-        if (user){
+        if (user) {
             return res.status(200).json(user)
-        }else{
+        } else {
             return res.status(404).json({
                 message: "aucun utilisateur trouvé"
             })
         }
     } catch (error) {
-        return res.status(500).json({ error})
+        return res.status(500).json({ message: error })
     }
 }
 
 //Modifier un utilisateur
-exports.ModifyUser = async (req, res) => {
+exports.modifyUser = async (req, res) => {
     try {
         const userObject = req.file ? {
             ...req.body,
-            // password: hash,
             profilepic: `${req.protocol}://${req.get('host')}/images/profile/${req.file.filename}`
         } : {
-            ...req.body,
-            // password: hash,
-        }
-        if(req.body.password){
-            const hash = await bcrypt.hash(req.body.password, 10)
-            userObject.password = hash
+            ...req.body
         }
         const user = await User.findOne({
             where: {
@@ -45,13 +39,19 @@ exports.ModifyUser = async (req, res) => {
             }
         })
         if (user) {
-            if (req.file) {
-                const toDelete = user.profilepic.split('/profile/')[1];
-                try {
-                    fs.unlinkSync(`images/profile/${toDelete}`)
-                } catch (error) {
-                    console.error(error)
+            if (req.body.password) {
+                const isValid = await bcrypt.compare(req.body.oldPassword, user.password)
+                if (!isValid) {
+                    return res.status(400).json(
+                        { message: "Ancien mot de passe incorrect" }
+                    );
                 }
+                const hash = await bcrypt.hash(req.body.password, 10)
+                userObject.password = hash
+            }
+            let toDelete
+            if (user.profilepic && req.file) {
+                toDelete = user.profilepic.split('/profile/')[1];
             }
             await User.update({
                 ...userObject
@@ -60,13 +60,20 @@ exports.ModifyUser = async (req, res) => {
                     id: req.params.id
                 }
             })
+            if (toDelete) {
+                fs.unlinkSync(`images/profile/${toDelete}`)
+            }
             return res.status(200).json({
                 message: ' Profile mis à jour'
             });
+        } else {
+            return res.status(404).json({
+                message: "Utilisateur non trouvé"
+            })
         }
     } catch (error) {
-        return res.status(400).json({
-            error: error
+        return res.status(500).json({
+            message: error
         });
     }
 }
@@ -79,14 +86,19 @@ exports.deleteUser = async (req, res) => {
                 id: req.params.id
             }
         })
-        const toDelete = user.profilepic.split('/profile/')[1];
         if (user) {
-            fs.unlinkSync(`/profile/${toDelete}`)
+            let toDelete
+            if (user.profilepic) {
+                toDelete = user.profilepic.split('/profile/')[1];
+            }
             await User.destroy({
                 where: {
                     id: req.params.id
                 }
             })
+            if (toDelete){
+                fs.unlinkSync(`/profile/${toDelete}`)
+            }
             return res.status(200).json({
                 message: 'Profile retiré'
             })
@@ -94,7 +106,7 @@ exports.deleteUser = async (req, res) => {
 
     } catch (error) {
         return res.status(400).json({
-            error: error
+            message: error
         });
     }
